major: admin routing fixed
