+++ src/app/admin/dashboard/page.jsx
@@ -1,0 +1,8 @@
+import Dashboard from "../../../components/Admin/Dashboard";
+import React from "react";
+
+const Page = () => {
+  return <Dashboard />;
+};
+
+export default Page;
