+++ src/components/Admin/Dashboard.jsx
@@ -3,8 +3,11 @@ import React from "react";
 import { Box, Grid } from "@mui/material";
 import StatsCards from "./StatsCards";
 import BookingsCard from "./BookingCard";
+import { useBookings } from "../../contexts/BookingsContext";
 
-const Dashboard = ({ bookings }) => {
+const Dashboard = ({}) => {
+  const { bookings } = useBookings();
+
   return (
     <Box sx={{ padding: "16px" }}>
       <StatsCards />
